requirements.txt
Pillow==9.2.0
pip==21.2.3
setuptools==57.4.0
